refactor: 파일 역할에 맞게 네이밍 변경
