+++ data_loaders/__init__.py
@@ -1,1 +1,1 @@
-from .script import *
+from .data_loader_factory import *
